common/bitmap: Don't fail on realloc (ptr, 0)
The following commands:

  nbdkit -fv --filter=cow memory size=512 --run 'qemu-img info $nbd'
  nbdkit -fv --filter=cache memory size=512 --run 'qemu-img info $nbd'
  nbdkit -fv --filter=cow null --run 'qemu-img info $nbd'

all fail with:

  nbdkit: memory[1]: error: realloc: Success

Initial git bisect pointed to commit 3166d2bcbfd2 (this is not real
cause, it merely exposes the bug).

The reason this happens is because the new behaviour after
commit 3166d2bcbfd2 is to round down the size of the underlying disk
to the cow/cache filter block size.

=> Size of the underlying disk is 512 or 0 bytes,
   block size is 4096 bytes.

=> Size of the disk is rounded down to 0.

=> The cow/cache filter requests a bitmap of size 0.

=> bitmap_resize calls realloc (ptr, 0).

=> The glibc implementation of realloc returns NULL + errno == 0.
   (Other realloc implementations can behave differently.)

=> This is not an error, but the existing code thinks it is because of
   the NULL return.

This commit changes the code so it doesn't bother to call realloc if
the new bitmap size would be 0.

There are many other places in nbdkit where we call realloc, and I did
not vet any of them to see if similar bugs could be present, but it is
quite likely.

Note in passing that the correct way to use the cow/cache filter with
a disk which isn't a multiple of the block size is to combine it with
the truncate filter, eg:

  nbdkit -fv --filter=cow --filter=truncate memory size=512 round-up=4096

Thanks: Eric Blake