server: Improve check of string validity
Factor out a common function for checking whether a client export name
is valid, checking for length and no embedded NULs (in the future, we
may also add a check for well-formed UTF-8).  Checking for embedded
NUL is important now that we allow plugins to alter content based on
the client's request: if the client requested 'a\0b' (against the NBD
protocol), they would be surprised if the reflection plugin only gave
back 'a'.  Rejecting bad strings is easier than altering our API to
pass through lengths to preserve embedded NUL.

Use the new functions to simplify existing code in NBD_OPT_EXPORT_NAME
and NBD_OPT_GO, but also to validate the export name and context query
handed to NBD_OPT_SET_META_CONTEXT.

No known existing client makes it easy to send 'a\0b' as an export or
meta context name against the protocol, but well-placed breakpoints in
gdb on a client process allowed me to test this manually.

Signed-off-by: Eric Blake <eblake@redhat.com>