common/protocol: Declare additional constants
We aren't using them yet, but as we are now publishing this file, it's
worth letting other clients have a chance to use named bits reserved
by the protocol.  This does not expose anything related to resize, as
that is still experimental.

Also tweak a formatting nit I noticed while scanning the list of #defines.

Signed-off-by: Eric Blake <eblake@redhat.com>