server: Switch to fixed-length conn->exportname
Since each connection already has a malloc()d struct conn, and the NBD
protocol places a fixed length on the maximum string at 4k, it's not
that much harder to just directly set aside 4k in the connection
struct instead of allocating a name on the fly.  In the common case of
short names, it does mean that a connection now occupies slightly more
unused heap space, but these days, 4k overhead is probably in the
noise.  Doing this makes it easier for an upcoming patch to refactor
name validation without having to worry about an allocation failure,
which in turn makes it easier to keep the connection alive even after
failing NBD_OPT_GO (where an allocation failure would have to be
handled differently than the diagnosis of an invalid client request).

This also points out a spot where we need to enlarge
MAX_OPTION_LENGTH, but that will be a separate patch.

Signed-off-by: Eric Blake <eblake@redhat.com>