retry: Check next_ops->can_FOO on retry
After a retry, if the second connection has fewer permissions than the
first, but we blindly call next_ops->FOO, we end up triggering an
assertion failure in backend.c.  This is particularly noticeable when
the retry_readonly flag is in effect, as that makes it much easier for
there to be fewer permissions than before.  Logging the number of
reopens helps show that it is not worth retrying a write-like command
after the first retry has changed the mode to readonly.

The next patch will add testsuite coverage of pwrite to demonstrate.

Signed-off-by: Eric Blake <eblake@redhat.com>