fuzzing: Increase memory limit to 256M (from 50M).
Since we added the static zero buffer (see server/plugins.c:
plugin_zero) the binary now requires more virtual memory to start up.
It's unclear if it requires real memory or if the buffer is still
being allocated lazily.  In any case it means we need to increase the
size of the memory limit used by AFL from the default.