server: Refactor to drop connection_get_handle()
Add backend_finalize() as one remaining function in backend.c, at
which point all backend functions that use a handle now also have a
wrapper.  As such, it's easier to change the signature to have the
backend hand the handle to filters.c or plugins.c, and drop the
connection_get_handle() function entirely.

Signed-off-by: Eric Blake <eblake@redhat.com>