retry: Don't call into closed plugin
If reopen fails, calling into the plugin would end up passing NULL as
the handle.  For the sh plugin, this merely results in a truncated
command line, but for other plugins it could result in a crash.  The
fix is to have the retry filter track whether reopen failed, and
short-circuit any new transaction to begin with a retry when the
previous transaction ended in that state.

Add asserts to backend to ensure that the plugin's handle is non-NULL
at all points of use, now that the retry filter keeps that contract.

Fixes: f0f0ec49
Signed-off-by: Eric Blake <eblake@redhat.com>