server: Allow file descriptors to be passed to nbdkit_read_password.
Allow password parameters such as ‘password=-FD’ where FD is a file
descriptor number inherited by nbdkit from the parent process.  This
is another way to allow programs to hand passwords to nbdkit in a very
secure way, for example over a pipe so they never touch the
filesystem.

Previously nbdkit allowed you to use literal passwords on the command
line if they began with a ‘-’ (but were not just that single
character).  However that was contrary to the documentation, and this
commit now prevents that.