vddk: Don't compile this plugin on non-x86 platforms.
VMware has only ever shipped VDDK on x86.  (I cannot find an
authoritative answer to whether it ever shipped on Itanium.)

For 32 bit (i686) it only existed up to VDDK 5.1.1, which is the very
minimum version of VDDK that we support.  However we can still compile
the plugin there.

Don't compile the plugin on non-x86 architectures.  Although it can be
compiled, the plugin is of no use on these architectures unless it can
be coupled with VDDK, and that does not exist.