server: Don't return connection status from handle_single_connection.
This was only used in one place: when using the -s option on the
command line.  In this case nbdkit -s would exit with EXIT_FAILURE if
there was some kind of I/O error or early client close.  This is not
particularly useful because: (1) It prevents the plugin being unloaded
properly. (2) Client close is a normal shutdown strategy for some
clients, so not really an error.  (3) It's undocumented and
inconsistent with other server modes.  (4) It's awkward when doing
fuzzing.

Note that connection_get_status (added in commit 0e1f158a1a
"connections: Add thread-safe status indicator") is still required to
coordinate connection shutdown across threads.