docs: Turn previous release notes for nbdkit 1.4 - 1.14 into man pages.
Release notes are important for the team who do Quality Assurance on
nbdkit and similar projects, since they use it to drive their test
development, so having nicely organized, correct release notes is
important for nbdkit.

In libguestfs we have an equivalent release notes man page but all
releases are placed in a single page which is very unmanageable.  To
avoid this mistake I intend to create a new release notes man page for
every stable release, and in future I will likely go an split up the
libguestfs release notes page too.

I don't think that there were release notes for nbdkit 1.2, as it was
a kind of "initial release" (albeit one which took place 5 years after
the project started) and so these release notes are missing.