tests: Refactor Makefile to split out TESTS variables.
Make it easier to add tests by making each test work more
independently of the others.  We add two new variables,
LIBGUESTFS_TESTS and GUESTFISH_TESTS to list tests which depend on
libguestfs C library and guestfish respectively.  These are added to
the TESTS variable at the end if the respective requirement is
available.

Note that GUESTFISH_TESTS is removed in the following commit, because
all of these tests are shell scripts and we'd prefer to have tests
check for guestfish at runtime using ‘requires’.  However doing it
this way makes it easier to understand the change.