python: Make robust against traceback.format_exception returning error.
This Python method can return an error in some instances, for example
in the case where the wrong number of parameters is passed to a plugin
function.  This would cause the Python plugin to segfault.  Check the
return value is not NULL to make the plugin robust against this.

(cherry picked from commit 4963c888955f70bb5f302582b927b9cc05d11a65)