python: Drop pointless NULL parameter to PyObject_CallFunction
A trailing NULL is required for PyObject_CallFunctionObjArgs() to
terminate a list of objects with no other indication of how many were
passed through varargs; but it is not needed when calling
PyObject_CallFunction() where the format string precisely indicates
how many other varargs to consume.  4 out of 5 uses had an extra
parameter; thankfully, it caused no visible misbehavior.

Fixes: 5ae45c00
Fixes: fd8cc6c8
Signed-off-by: Eric Blake <eblake@redhat.com>