Implement .preconnect callback.
This is called just after TCP connection but before any NBD
negotiation or TLS authentication.  It can be used for early
whitelisting of connections or, with more care, for rate limiting of
connections.  You have to be careful of denial of service attacks
here.

The callback can also be filtered in filters.

This is so far only available from the C API.