sh: Remove tmpdir and script globals from call.c and call.h.
tmpdir was not used at all.  The script global was used only for
printing error messages, but call.c can get this (and more easily and
more reliably too) just by looking at argv[0] from the array of
arguments it is passed.

This is neutral refactoring but helps to decouple sh.c from call.c.