plugins: Rename *_old slots in struct nbdkit_plugin to *_v1.
When API v2 is requested, and also internal to the server, the slots
used by old pread (etc) functions were called _pread_old (etc).  Since
we plan eventually to deprecate the v2 functions this would lead to
awkward renaming (_pread_old to _pread_old_old), and the _old suffix
is not descriptive of what the function signifies.  Instead name them
_pread_v1 (etc).