docs, tests: Improved method to probe for plugins and filters.
Previously we suggested using nbdkit --dump-plugin to probe for
plugins and filters, but this could fail if a plugin exists but the
dump_plugin callback is broken.  In any case it's simpler to
use --version for probing and the output is more informative.  This
works even with very old versions of nbdkit (I checked this back to
nbdkit 1.2).

This also adds two tests to ensure this method does not regress in
future.