ext2: Replace bad advice with TODO
Our advice to apply the partition filter to the ext2 plugin is flawed
- that tries to treat the raw disk as a file system, extract a file,
then read a partition header from that file to select a subset of that
file.  What we really need is an ext2 filter, so we can take a raw
disk, extract the partition containing an ext2 filesystem, then
extract the file.  But to do that, we need an ext2 filter, where we
handwrite our own io_manager that calls into our backend code, rather
than using the unix_io_manager that comes with ext2fs.

Signed-off-by: Eric Blake <eblake@redhat.com>