filters: More on next_ops(nxdata) stability
It turns out that it's not just enough to have a stable nxdata, but
also a stable next_ops.  Of course, since our next_ops is a static
struct populated only with function pointers and nothing that changes
dynamically (unlike the nxdata which DOES change per connection), we
already meet this condition; but document it to make sure we keep our
promise.

Also, adding testsuite coverage never hurts ;)

Signed-off-by: Eric Blake <eblake@redhat.com>