filters: Add ext2 filter
The ext2 plugin does not work with partitions, and requires the disk
image to be in the local filesystem.  Better is to have a filter that
lets any other plugin be the disk image (including with the partition
filter), where the ext2 code then calls into our backend code.  But
since our backend code is different than Unix fd I/O, we need a new
custom io_manager (I heavily borrowed ideas from the existing
unix_io_manager and test_io_manager in e2fsprog).

With that in place, the new filter version of ext2.c is very similar
to the plugin version: it only needs tweaks for the different filter
API, and for calling into the new io_manager with the 'const char
*name' parameter actually being the opaque parameter that lets the
rest of io.c call back into nbdkit.

The next commit will then deprecate the ext2 plugin.

Signed-off-by: Eric Blake <eblake@redhat.com>