ext2: Add mode for letting client exportname choose file from image
Not enabled by default, because of the potential security concern that
a client can use this to sniff what file names exist in the image; but
for an opt-in setting in the server, this lets a single nbdkit process
serve multiple files from an ext2 image (one at a time, unless we find
a way to allow a filter to open the plugin once without a current
connection, then share that opened plugin across multiple client
connections).

In testing this manually, I learned that qemu 4.2 fails to preserve
leading / in nbd://host:port//path/to/file (no matter how many slashes
you add), so the code here adds a leading slash to be nice to clients
like broken qemu.

Signed-off-by: Eric Blake <eblake@redhat.com>