server: filters: Remove struct b_h.
This was previously used as ‘nxdata’ and stored a tuple of ’b->next’
and the real filter handle.  However after recent changes we don't
need it.  We can use ‘b->next’ as nxdata, and the handle is passed to
us by the calling functions.

Inspired by Eric Blake’s observations in this email:
https://www.redhat.com/archives/libguestfs/2020-February/msg00092.html