filters: Remove most next_* wrappers
With our recent cleanups to nxdata, the only remaining difference
between functions like backend_open() and next_open() was the
signature (the latter used void*, the former struct backend *); the
ABI is compatible.  All of our filters are in-tree, and we don't
promise API/ABI stability, but it is still a lot of files to touch if
we change the public signature, so the simplest solution to avoid the
redundant hop through wrapper functions is to change our header to
compile with a typedef, which is either 'struct backend *' for
internal code or 'void *' for filters.  With this in place, we can now
delete most of the next_* wrappers, by pointing to corresponding
backend_* functions instead; the few exceptions are .config and
.config_complete (the public API returns an integer, but the backend
code exits on failure and thus returns void), and .preconnect (we
don't have a backend_preconnect function).

Signed-off-by: Eric Blake <eblake@redhat.com>