main: Restore argv after .config
The Linux kernel's /proc/NNN/cmdline is not a static copy of the
arguments passed in to exec*, but rather a dynamic view of the region
of userspace memory in the process.  Although you can modify argv[] at
runtime, those modifications are reflected through to /proc/NNN, which
can be confusing for determining what a long-running nbdkit process is
doing.  Hence, although we can make temporary modifications if it
makes .config easier, we should undo those modifications for the long
term.

For example, with 'nbdkit --filter=truncate memory 1 truncate=2', this
patch changes /proc/NNN/cmdline from ending in what looks like two
arguments 'truncate' '2' back to the one-argument form 'truncate=2'.

Signed-off-by: Eric Blake <eblake@redhat.com>