vddk: Drop support for VDDK 5.1.1
VDDK 5.1.1 was the last version from VMware that supported 32-bit
platforms, but it was last updated in 2014 and it appears that VMware
no longer supports it.  It is also getting harder to find users that
care about this version.  The next oldest version, VDDK 5.5.5, dropped
support for 32-bit platforms, and we can slightly simplify our code by
documenting a larger minimum supported version.

Signed-off-by: Eric Blake <eblake@redhat.com>