linuxdisk: Compile this plugin conditionally if we don't have mke2fs -d.
This plugin cannot work by design unless we have mke2fs with a
working -d option.  On RHEL 7, mke2fs exists but it predates the -d
option.  On *BSD, mke2fs is not normally available by default.

The tests which use the ‘disk’ file (or its derivatives: ‘ext2.img’,
‘disk.gz’ and ‘disk.xz’) must also be skipped because the base file is
generated using this plugin.