server: Add nbdkit_shutdown() call.
Plugins and filters may call this to initiate an asynchronous shutdown
of the server.  This would only be used in the connected phase —
plugins should still call exit(3) directly for configuration failure.

It is equivalent to sending a kill signal to self, but it's cleaner to
have an API for this and better for potential future non-POSIX platforms.