server: Only display finalize/close debug messages if callback is called.
For example if .preconnect fails, it printed the debug messages:

  debug: preconnect: ...
  debug: close: ...

This confused me into thinking that the close callback was actually
being called without a corresponding open.  In fact it is not called
in this case.

Suppress the debug messages for finalize and close unless the callback
is actually being called.