tests: Don't let test-parallel-* hang on nbdkit bug
If nbdkit has a bug (such as the nbd-standalone bug fixed in the
previous commit), qemu-io ends up waiting forever rather than
realizing that if the server disappears unexpectedly then qemu-io
should quit.  So add timeouts so the testsuite will flag the problem
instead of hang (tested by reordering this commit before the
previous).

It's trickier than I expected: from the command line, 'timeout 10s
qemu-io ...' works just fine, but from a script, it fails.  Why?
Because timeout defaults to creating a new process group when run
non-interactively, but qemu-io is generally an interactive program
that expects to be able to modify the terminal settings when stdin is
a terminal, even when it will not be reading from stdin because of the
use of -c.  When stdin is inherited but no longer the controlling
terminal because of timeout's new process group, qemu-io fails to make
progress. As solutions, we can either use 'timeout --foreground', or
redirect stdin to something that is not a terminal; this patch uses
the latter.  Since the timeout is a last-ditch effort that won't
trigger in normal successful usage, we can make it long: 30s for
probing qemu-io, and 60s for the actual tests (when io is slower due
to our use of --filter=delay).

Signed-off-by: Eric Blake <eblake@redhat.com>