nbd: Normalize return values of can_*
Although nbdkit documents that any positive value should be treated as
success to the .can_* callbacks, we had a window of releases where
anything other than 1 could trigger an assertion failure, fixed in the
previous patch.  Update what we return to avoid tripping the bug in
broken nbdkit.

Our return value has been latent since 0e8e8eb11d (v1.1.17), but the
assertion failures did not occur until c306fa93ab and neighboring
commits (v1.15.1).  As v1.13.6 was the point where we started
preferring builds against libnbd, where we always returned 1 on
success, the problem was not detected until now; but it is still in
the wild for any user mixing old plugins with new libnbd.

Reported-by: Svenne Krap (https://svenne.dk)
Signed-off-by: Eric Blake <eblake@redhat.com>